combine typing
